move menu to page level
