+++ frontend/pages/index.js
@@ -2,7 +2,13 @@ import Layout from "../components/Layout.js";
 import React, { Component } from "react";
 import fetch from "isomorphic-unfetch";
 import Link from "next/link";
+import Menu from "../components/Menu.js";
 import { Config } from "../config.js";
+
+const headerImageStyle = {
+    marginTop: 50,
+    marginBottom: 50
+};
 
 class Index extends Component {
     static async getInitialProps(context) {
@@ -52,6 +58,12 @@ class Index extends Component {
         });
         return (
             <Layout>
+                <Menu location="header-menu" />
+                <img
+                    src="/static/images/wordpress-plus-react-header.png"
+                    width="815"
+                    style={headerImageStyle}
+                />
                 <h1>{this.props.page.title.rendered}</h1>
                 <div
                     dangerouslySetInnerHTML={{
